requirements.txt
streamlit
pandas

langchain
langchain_community
langchain-core
langgraph

langchain_openai

tavily-python

scikit-learn
altair



